signature-regn-api/Identifi_Code_Streamlit/requirements.txt
streamlit==1.28.1
requests==2.31.0
pillow==10.1.0
pandas==2.1.3
numpy==1.24.3
